requirement.txt
requests
python-Wappalyzer